finished task 3
wrote some new articles and tried to add a photo to one but it's going to need to be resized and I'd like to figure out how to change how far down each article expands dynamically according to how long the article is. I am not sure how to do it but I will at least look into it.